Move location of "GetLockPeriodInBlocksFromFormattedStringSpecifier" to work around build system issues.
